requirements.txt
fastapi==0.115.11
SQLALchemy==2.0.38
alembic==1.15.1
pydantic-settings==2.8.1
asyncpg==0.30.0